+++ client/src/features/Rush/BalanceGame/Background.tsx
