+++ client/src/features/Rush/BalanceGame/BalanceGameCard/RushCardResultDescription.tsx
