+++ client/src/features/Rush/BalanceGame/BalanceGameSection/FinalResult.tsx
