+++ client/src/features/Rush/BalanceGame/BalanceGameSection/SelectedCard.tsx
@@ -2,9 +2,9 @@ import { useState } from "react";
 import { motion } from "framer-motion";
 import { CARD_COLOR, CARD_DAYS, CARD_TYPE } from "@/constants/Rush/rushCard.ts";
 import { ASCEND, DISSOLVE, SCROLL_MOTION } from "@/constants/animation.ts";
-import RushCardCurrentRatio from "@/features/Rush/BalanceGameCard/RushCardCurrentRatio.tsx";
-import RushCardResultDescription from "@/features/Rush/BalanceGameCard/RushCardResultDescription.tsx";
-import RushCountDown from "@/features/Rush/BalanceGameCard/RushCountDown.tsx";
+import RushCardCurrentRatio from "@/features/Rush/BalanceGame/BalanceGameCard/RushCardCurrentRatio.tsx";
+import RushCardResultDescription from "@/features/Rush/BalanceGame/BalanceGameCard/RushCardResultDescription.tsx";
+import RushCountDown from "@/features/Rush/BalanceGame/BalanceGameCard/RushCountDown.tsx";
 import useCountDown from "@/hooks/useCountDown.ts";
 import ArrowLeftIcon from "/public/assets/icons/arrow-line-left.svg?react";
 import ArrowRightIcon from "/public/assets/icons/arrow-line-right.svg?react";
