+++ client/src/features/Rush/Casper/CasperDescription.tsx
@@ -1,7 +1,7 @@
 import { ReactNode } from "react";
 import { motion } from "framer-motion";
 import { ASCEND, SCROLL_MOTION } from "@/constants/animation.ts";
-import { CasperSubDescription } from "@/features/Rush/CasperSubDescription.tsx";
+import { CasperSubDescription } from "@/features/Rush/Casper/CasperSubDescription.tsx";
 
 interface CasperDescriptionProps {
     title: ReactNode;
