+++ client/src/features/Rush/Electric/ElectricReason.tsx
@@ -1,8 +1,8 @@
 import { useState } from "react";
 import Toggle from "@/components/Toggle";
 import { CARD_DATA, TOGGLE_OPTIONS } from "@/constants/Rush/electricCardData.tsx";
-import { ElectricReasonCard } from "@/features/Rush/ElectricReasonCard.tsx";
-import { ElectricSection } from "@/features/Rush/ElectricSection.tsx";
+import { ElectricReasonCard } from "@/features/Rush/Electric/ElectricReasonCard.tsx";
+import { ElectricSection } from "@/features/Rush/Electric/ElectricSection.tsx";
 import { SectionKeyProps } from "@/types/sections.ts";
 
 export function ElectricReason({ id }: SectionKeyProps) {
