+++ client/src/features/Rush/Electric/ElectricSection.tsx
