+++ client/src/features/Rush/Extra/FAQ.tsx
