+++ client/src/features/Rush/Extra/Intro.tsx
