+++ client/src/features/Rush/index.tsx
@@ -1,13 +1,13 @@
-export { CasperComfortable } from "./CasperComfortable.tsx";
-export { CasperFar } from "./CasperFar.tsx";
-export { CasperFast } from "./CasperFast.tsx";
-export { CasperSmartKey } from "./CasperSmartKey.tsx";
-export { CasperWide } from "./CasperWide.tsx";
-export { ElectricAdvantage } from "./ElectricAdvantage.tsx";
-export { ElectricReason } from "./ElectricReason.tsx";
-export { FAQ } from "./FAQ.tsx";
-export { Intro } from "./Intro.tsx";
-export { ReasonFirst } from "./ReasonFirst.tsx";
-export { ReasonSecond } from "./ReasonSecond.tsx";
-export { CasperCharge } from "./CasperCharge.tsx";
-export { BalanceGame } from "./BalanceGame.tsx";
+export { CasperComfortable } from "./Casper/CasperComfortable.tsx";
+export { CasperFar } from "./Casper/CasperFar.tsx";
+export { CasperFast } from "./Casper/CasperFast.tsx";
+export { CasperSmartKey } from "./Casper/CasperSmartKey.tsx";
+export { CasperWide } from "./Casper/CasperWide.tsx";
+export { ElectricAdvantage } from "./Electric/ElectricAdvantage.tsx";
+export { ElectricReason } from "./Electric/ElectricReason.tsx";
+export { FAQ } from "./Extra/FAQ.tsx";
+export { Intro } from "./Extra/Intro.tsx";
+export { ReasonFirst } from "./Reason/ReasonFirst.tsx";
+export { ReasonSecond } from "./Reason/ReasonSecond.tsx";
+export { CasperCharge } from "./Casper/CasperCharge.tsx";
+export { BalanceGame } from "./BalanceGame/BalanceGame.tsx";
